+++ client/src/routes/home-page/HomePage.jsx
@@ -1,4 +1,4 @@
 export const HomePage = () => {
-  // TODO: visualise courts
+  // TODO: visualise clubs by city + filters
   return <h1>This is the homepage</h1>;
 };
